# docs/openapi.yaml
# =============================================================================
# OpenAPI-спека digest-канваса.
#
# Источники правды (НЕ выдумывать поля и флоу сверх них):
#   - docs/user-scenarios.md   — US-1, US-2, US-3, US-6
#   - docs/message-contract.md — контракт Message
#
# Карта сценариев → эндпоинты:
#   US-1 (веб)            → POST /messages
#   US-2 (чат + @bot)     → POST /ingest/chat   (дергает бот, не фронт)
#   US-3 (личка боту)     → POST /ingest/dm     (дергает бот, не фронт)
#   US-6 (фильтры)        → GET /messages?dateFrom&dateTo&text&imageObject
#   треды (контракт)      → POST /messages с parentMessageId
#   канвасы               → /canvases  (сообщение всегда на одном canvasId)
# =============================================================================
openapi: 3.0.3

info:
  title: Digest Canvas API
  version: 0.2.0
  license:
    name: MIT
    url: https://opensource.org/license/mit
  description: |
    Digest — лента сообщений с трёх каналов: веб, групповой чат (тег бота),
    личка бота. Сообщения живут на канвасах. Треды — как вложенные блоки
    Obsidian: parentMessageId выбирается явно, не угадывается.

    Кто ходит куда:
    - фронт → `/canvases*`, `/messages*` (кука/сессия пользователя);
    - бот  → `/ingest/*` (свой токен, см. securitySchemes.botToken).

servers:
  - url: /api/v1
    description: Относительный префикс — фронт и бот ходят на тот же хост.

security:
  - session: []

tags:
  - name: canvases
    description: Канвасы. Сообщение принадлежит ровно одному канвасу.
  - name: messages
    description: Digest, создание сообщений, фильтры (US-1, US-6)
  - name: ingest
    description: Входящие от бота — чат и личка (US-2, US-3)

paths:
  /canvases:
    get:
      tags: [canvases]
      operationId: listCanvases
      summary: Канвасы текущего пользователя
      responses:
        "200":
          description: Список, новые сверху.
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/Canvas"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags: [canvases]
      operationId: createCanvas
      summary: Создать канвас
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CanvasCreate"
      responses:
        "201":
          description: Создан.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Canvas"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"

  /canvases/{canvasId}:
    get:
      tags: [canvases]
      operationId: getCanvas
      summary: Один канвас
      parameters:
        - $ref: "#/components/parameters/CanvasId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Canvas"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

  /messages:
    get:
      tags: [messages]
      operationId: listMessages
      summary: Лента digest с фильтрами (US-6)
      description: |
        Все фильтры объединяются по AND. Пустые фильтры = вся лента канваса.
        Сброс фильтров на фронте = запрос без query (кроме обязательного canvasId).

        Отложенные сообщения (dateTimeReveal в будущем): всем, кроме автора,
        отдаём маску — text/media вырезаны, остальные поля на месте, чтобы
        фронт мог показать плейсхолдер «откроется в …».
      parameters:
        - name: canvasId
          in: query
          required: true
          description: Лента всегда в рамках одного канваса.
          schema:
            type: string
        - name: dateFrom
          in: query
          description: Нижняя граница по dateTimeSend, включительно.
          schema:
            type: string
            format: date-time
        - name: dateTo
          in: query
          description: Верхняя граница по dateTimeSend, включительно.
          schema:
            type: string
            format: date-time
        - name: text
          in: query
          description: |
            Подстрока без учёта регистра. Ищем в теле сообщения и в его
            комментах (детях) — если совпал комментарий, в выдачу попадает
            родитель (см. US-6).
          schema:
            type: string
        - name: imageObject
          in: query
          description: |
            Фильтр «объект на картинке»: «кот», «чек», «паспорт».
            Нет картинок у сообщения → мимо. Есть объект → попадает.
          schema:
            type: string
      responses:
        "200":
          description: Отфильтрованная лента, сортировка по dateTimeSend по убыванию.
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/Message"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: canvasId не существует.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    post:
      tags: [messages]
      operationId: createMessage
      summary: Создать сообщение из веба (US-1)
      description: |
        Коммент/reply — то же Message с заполненным parentMessageId
        (см. message-contract.md). Канал = web.

        Сервер сам проставляет id, dateTimeSend, senderId (из сессии) —
        клиент их не шлёт, пришлёт — 400.

        Правила parentMessageId (Obsidian-style):
        - null / нет поля  → корень канваса;
        - id карточки      → ребёнок ровно той карточки, что выбрал юзер;
        - родитель обязан существовать и жить на том же canvasId;
        - циклы невозможны by construction (родитель уже существует).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageCreate"
      responses:
        "201":
          description: Создано. Возвращаем полный объект — фронт дорисовывает ленту без refetch.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "400":
          description: |
            Пустое тело (нет ни text, ни media), кривой mime, dateTimeReveal
            раньше dateTimeSend, лишние серверные поля.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: parentMessageId или canvasId не существует (или родитель с другого канваса).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /messages/{messageId}:
    get:
      tags: [messages]
      operationId: getMessage
      summary: Одно сообщение (карточка)
      parameters:
        - $ref: "#/components/parameters/MessageId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

  /ingest/chat:
    post:
      tags: [ingest]
      operationId: ingestChatMessage
      summary: Забрать сообщение из группового чата (US-2)
      description: |
        Бот дергает это, когда его тегнули (@bot …) или сделали reply на него.
        Без тега бот сюда не ходит — фильтрация на стороне бота.

        Дедупликация: (chatId, sourceMessageId) уникальны. Повторный тег того
        же сообщения → 200 с уже существующим Message, а не дубликат (см. US-2).
      security:
        - botToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatIngest"
      responses:
        "201":
          description: Новое сообщение в digest, канал = chat.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "200":
          description: Уже собирали (повторный тег) — возвращаем существующее.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: canvasId не существует.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /ingest/dm:
    post:
      tags: [ingest]
      operationId: ingestDirectMessage
      summary: Забрать сообщение из лички бота (US-3)
      description: |
        Любое непустое входящее в личку → в digest, канал = dm.
        Команды бота (/start и прочее) сюда не шлём — бот отсеивает их сам.
      security:
        - botToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DmIngest"
      responses:
        "201":
          description: В digest. Бот после этого подтверждает приём юзеру.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: canvasId или parentMessageId не существует.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

components:
  securitySchemes:
    session:
      type: apiKey
      in: cookie
      name: sid
      description: Сессионная кука юзера. Из неё сервер берёт senderId / ownerId.
    botToken:
      type: http
      scheme: bearer
      description: Токен бота для /ingest/*. Пользовательские ручки живут на сессии.

  parameters:
    MessageId:
      name: messageId
      in: path
      required: true
      schema:
        type: string
    CanvasId:
      name: canvasId
      in: path
      required: true
      schema:
        type: string

  responses:
    BadRequest:
      description: Невалидный запрос.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Ресурс не найден.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Нет сессии или нет/протух botToken.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"

  schemas:
    Canvas:
      type: object
      required: [id, title, ownerId, createdAt]
      properties:
        id:
          type: string
          readOnly: true
        title:
          type: string
        ownerId:
          type: string
          readOnly: true
          description: Кто создал. Сервер берёт из сессии.
        createdAt:
          type: string
          format: date-time
          readOnly: true

    CanvasCreate:
      type: object
      required: [title]
      properties:
        title:
          type: string
          minLength: 1
          description: Трим на бэке. Пустая строка после трима → 400.

    Message:
      type: object
      description: 1:1 с docs/message-contract.md. channel ставит сервер (US-1/2/3).
      required: [id, canvasId, dateTimeSend, senderId, channel]
      properties:
        id:
          type: string
          readOnly: true
          description: Выдаёт система. Клиент при создании не шлёт.
        text:
          $ref: "#/components/schemas/TextBlock"
        revealAnimation:
          type: string
          nullable: true
          description: Имя анимации появления. Нет значения = без анимации.
        media:
          type: array
          description: Вложения. Пустой массив = нет медиа.
          items:
            $ref: "#/components/schemas/Media"
        dateTimeSend:
          type: string
          format: date-time
          readOnly: true
          description: Время отправки. Ставит сервер. Не путать с reveal.
        canvasId:
          type: string
          description: Канвас-владелец. Перенос на другой канвас — отдельная операция.
        dateTimeReveal:
          type: string
          format: date-time
          nullable: true
          description: |
            Отложенный показ. null/нет = показать сразу.
            Должно быть >= dateTimeSend. До наступления все, кроме автора,
            видят маску без text/media.
        senderId:
          type: string
          readOnly: true
          description: Автор. Сервер берёт из сессии или из ingest-пейлоада.
        parentMessageId:
          type: string
          nullable: true
          description: |
            Треды как в Obsidian. null = корень канваса.
            Выбирается явно (reply/коммент), никогда не угадывается.
            Родитель — тот же canvasId, циклы запрещены.
            Удалили родителя → дети становятся корнями, с канваса не пропадают.
        canvasPosition:
          $ref: "#/components/schemas/CanvasPosition"
        tag:
          type: string
          nullable: true
          description: Один тег. Пусто = без тега.
        channel:
          type: string
          readOnly: true
          enum: [web, chat, dm]
          description: Откуда пришло. web=US-1, chat=US-2, dm=US-3. Ставит сервер.

    MessageCreate:
      type: object
      description: |
        Клиент шлёт всё, кроме id / dateTimeSend / senderId — их ставит сервер.
        Пустое тело запрещено — нужен text и/или хотя бы один media.
      anyOf:
        - required: [text]
        - required: [media]
      required: [canvasId]
      properties:
        text:
          $ref: "#/components/schemas/TextBlock"
        revealAnimation:
          type: string
        media:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/Media"
        canvasId:
          type: string
        dateTimeReveal:
          type: string
          format: date-time
          description: Для отложенных. Раньше dateTimeSend → 400.
        parentMessageId:
          type: string
          description: Заполнено = коммент/reply. Пусто = корень.
        canvasPosition:
          $ref: "#/components/schemas/CanvasPosition"
        tag:
          type: string

    TextBlock:
      type: object
      description: Текст + опциональный стиль. Стиль без body не бывает.
      required: [body]
      properties:
        body:
          type: string
          minLength: 1
          description: Трим на бэке. Пустая строка после трима = текста нет.
        fontType:
          type: string
        design:
          type: string
        size:
          type: number
          description: Кегль в px.

    Media:
      type: object
      required: [mime, url]
      properties:
        mime:
          type: string
          enum:
            - video/mp4
            - audio/mpeg
            - image/jpeg
            - image/png
            - image/gif
        url:
          type: string
          format: uri

    CanvasPosition:
      type: object
      description: |
        Координаты карточки. Не прислали — сервер кладёт рядом с родителем
        или в дефолтную точку. Задано явно = ручной якорь.
      required: [x, y]
      properties:
        x:
          type: number
        y:
          type: number

    ChatIngest:
      type: object
      description: |
        US-2. Бот уже проверил, что его тегнули. Автор digest-сообщения =
        автор исходной реплики (authorId).
      required: [chatId, sourceMessageId, authorId, taggedById, canvasId]
      anyOf:
        - required: [text]
        - required: [media]
      properties:
        chatId:
          type: string
        chatTitle:
          type: string
        sourceMessageId:
          type: string
          description: Id исходного сообщения в чате. Пара (chatId, sourceMessageId) = ключ дедупа.
        authorId:
          type: string
          description: Автор исходной реплики. Станет senderId.
        taggedById:
          type: string
          description: Кто позвал бота.
        canvasId:
          type: string
        text:
          $ref: "#/components/schemas/TextBlock"
        media:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/Media"
        sourceUrl:
          type: string
          format: uri
        parentMessageId:
          type: string
          description: Reply на сохранённое сообщение в digest.

    DmIngest:
      type: object
      description: |
        US-3. Личка боту. Команды (/start и т.п.) бот сюда не пересылает.
        parentMessageId возможен, если юзер в личке делает reply на карточку.
      required: [authorId, sourceMessageId, canvasId]
      anyOf:
        - required: [text]
        - required: [media]
      properties:
        authorId:
          type: string
          description: Отправитель. Станет senderId.
        sourceMessageId:
          type: string
          description: Id сообщения в личке — ключ дедупа вместе с authorId.
        canvasId:
          type: string
        text:
          $ref: "#/components/schemas/TextBlock"
        media:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/Media"
        parentMessageId:
          type: string

    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: Машиночитаемый код (empty_body, bad_mime, parent_not_found, reveal_in_past…).
        message:
          type: string
          description: Человекочитаемо, можно показать в тосте.